Edit Validation To Real Time Validation & Add Feature To Delete Function

+++ js/script.js
@@ -4,6 +4,13 @@ var productCategory = document.getElementById("productCategory")
 var productDesc = document.getElementById("productDesc")
 var addBtn = document.getElementById("addBtn")
 var errorsAlert = document.getElementById("errorsAlert")
+
+let alert1 =document.getElementById("alert1")
+let alert2 =document.getElementById("alert2")
+let alert3 =document.getElementById("alert3")
+let alert4 =document.getElementById("alert4")
+
+
 var x ; // Index Of Object You Want To Update
 var errors =``
 
@@ -37,12 +44,11 @@ function addProduct()
             localStorage.setItem("productlist",JSON.stringify(productContainer))  
             clearproduct()
             display ()
-            errorsAlert.style="display:none;"
+           ResetInputs ()
         }
         else
         {
-            errorsAlert.style="display:inline-block;"
-            errorsAlert.innerHTML=errors
+            wrongAlert ()
         }
 
     }
@@ -60,15 +66,44 @@ function updateProduct()
      localStorage.setItem("productlist",JSON.stringify(productContainer))  
     display ()
     clearproduct()
+    ResetInputs ()
     addBtn.innerHTML="AddProduct"
-    errorsAlert.style="display:none;"
     }
     
    else
    {
-   errorsAlert.style="display:inline-block;"
-   errorsAlert.innerHTML=errors
+    wrongAlert ()
     }
+}
+function ResetInputs ()
+{  
+    productName.classList.remove("is-valid")
+    productPrice.classList.remove("is-valid")
+    productCategory.classList.remove("is-valid")
+    productDesc.classList.remove("is-valid")
+    productName.classList.remove("is-invalid")
+    productPrice.classList.remove("is-invalid")
+    productCategory.classList.remove("is-invalid")
+    productDesc.classList.remove("is-invalid")
+    errorsAlert.classList.replace("d-block","d-none")
+    alert1.classList.replace("d-block","d-none")
+    alert2.classList.replace("d-block","d-none")
+    alert3.classList.replace("d-block","d-none")
+    alert4.classList.replace("d-block","d-none")
+}
+function wrongAlert ()
+{
+    errorsAlert.innerHTML=errors;
+    errorsAlert.classList.replace("d-none","d-block")
+    alert1.classList.replace("d-block","d-none")
+    alert2.classList.replace("d-block","d-none")
+    alert3.classList.replace("d-block","d-none")
+    alert4.classList.replace("d-block","d-none")
+    
+   
+     // nameInput.classList.remove("is-invalid")
+    // emailInput.classList.remove("is-invalid")
+    // passwordInput.classList.remove("is-invalid")
 }
 
 function clearproduct()
@@ -91,7 +126,7 @@ function display ()
         <th>${productContainer[i].price}</th>
         <th>${productContainer[i].cate}</th>
         <th>${productContainer[i].desc}</th>
-        <th><button onclick="showUpdateProducts(${i})" class="btn btn-outline-warning">Update</button></th>
+        <th><button onclick="showUpdateProducts(${i})" class="btn btn-warning">Update</button></th>
         <th><button onclick="deleteEle(${i})" class="btn btn-outline-Danger">Delete</button></th> 
     </tr>`
     }
@@ -100,11 +135,15 @@ function display ()
 
 function deleteEle(index)
 {
-productContainer.splice(index,1)
-localStorage.setItem("productlist",JSON.stringify(productContainer))
-display ()
-addBtn.innerHTML="AddProduct"
-clearproduct()
+    if(window.confirm("Are You Sure You Want To Delete This Product ?")==true)
+    {
+        productContainer.splice(index,1)
+        localStorage.setItem("productlist",JSON.stringify(productContainer))
+        display ()
+        addBtn.innerHTML="AddProduct"
+        // clearproduct()
+    }
+
 }
 
 function showUpdateProducts(index)
@@ -115,6 +154,8 @@ function showUpdateProducts(index)
     productDesc.value=productContainer[index].desc
     addBtn.innerHTML="Update"
     x=index 
+    ResetInputs ()
+ 
 }
 
 function search(searchTerm)
@@ -124,7 +165,9 @@ function search(searchTerm)
     for(var i = 0 ; i<productContainer.length ;i++)
     {
         if(productContainer[i].name.toLowerCase().includes(searchTerm.toLowerCase())==true
-        ||productContainer[i].cate.toLowerCase().includes(searchTerm.toLowerCase())==true )
+        ||productContainer[i].cate.toLowerCase().includes(searchTerm.toLowerCase())==true
+        ||productContainer[i].desc.toLowerCase().includes(searchTerm.toLowerCase())==true
+        ||productContainer[i].price.toLowerCase().includes(searchTerm.toLowerCase())==true )
         {
             cartona+=`<tr>
             <th id="demo">${i}</th>
@@ -132,7 +175,7 @@ function search(searchTerm)
             <th>${productContainer[i].price}</th>
             <th>${productContainer[i].cate}</th>
             <th>${productContainer[i].desc}</th>
-            <th><button class="btn btn-outline-warning">Update</button></th>
+            <th><button class="btn btn-warning">Update</button></th>
             <th><button onclick="deleteEle(${i})" class="btn btn-outline-Danger">Delete</button></th> 
         </tr>`
         }
@@ -144,56 +187,92 @@ function search(searchTerm)
     document.getElementById("tableBody").innerHTML=cartona
 }
 
-
+productName.addEventListener("keyup",validateNameInput)
 function validateNameInput()
 {
+    console.log("fd")
     var regex = /^[A-Z][ a-zA-Z]{2,10}$/
     if (regex.test(productName.value)==true)
     {
+        productName.classList.add("is-valid")
+        productName.classList.remove("is-invalid")
+        alert1.classList.replace("d-block","d-none")
         return true
     }       
     else
     {
-        errors += `Enter Valid Name (First Letter : Capital - Then:From 2 To 10 Any Capital Or Small Letters)`
+        productName.classList.add("is-invalid")
+        productName.classList.remove("is-valid")
+        alert1.classList.add("d-block")
+        alert1.innerHTML=`Enter Valid Name (First Letter : Capital - Then:From 2 To 10 Any Capital Or Small Letters)`
+
+        errors += `Enter Valid Name`
         return false
     }
 }
+productPrice.addEventListener("keyup",validatePriceInput)
 function validatePriceInput()
 {
     var regex = /^([1-9][0-9][0-9][0-9]?|10000)$/ //100-10000
     if (regex.test(productPrice.value)==true)
     {
+        productPrice.classList.add("is-valid")
+        productPrice.classList.remove("is-invalid")
+        alert2.classList.replace("d-block","d-none")
         return true
     }       
     else
     {
-        errors += `Enter Valid Price (Numbers From:100 - To:10000 )`
+        productPrice.classList.add("is-invalid")
+        productPrice.classList.remove("is-valid")
+        alert2.classList.add("d-block")
+        alert2.innerHTML= `Enter Valid Price (Numbers From:100 - To:10000 )`
+
+        errors += `Enter Valid Price `
         return false
     }
 }
+productCategory.addEventListener("keyup",validateCategoryInput)
 function validateCategoryInput()
 {
     var regex = /^[A-Z][ a-zA-Z]{1,10}$/ 
     if (regex.test(productCategory.value)==true)
     {
+        productCategory.classList.add("is-valid")
+        productCategory.classList.remove("is-invalid")
+        alert3.classList.replace("d-block","d-none")
         return true
     }       
     else
     {
-        errors += `Enter Valid Category (First Letter : Capital - Then:From 1 To 10 Any Capital Or Small Letters)`
+        productCategory.classList.add("is-invalid")
+        productCategory.classList.remove("is-valid")
+        alert3.classList.add("d-block")
+        alert3.innerHTML= `Enter Valid Category (First Letter : Capital - Then:From 1 To 10 Any Capital Or Small Letters)`
+
+        errors += `Enter Valid Category`
         return false
     }
 }
+productDesc.addEventListener("keyup",validateDescInput)
 function validateDescInput()
 {
     var regex = /^[A-Z][ a-zA-Z]{2,20}$/ 
     if (regex.test(productDesc.value)==true)
     {
+        productDesc.classList.add("is-valid")
+        productDesc.classList.remove("is-invalid")
+        alert4.classList.replace("d-block","d-none")
         return true
     }       
     else
     {
-        errors += `Enter Valid Description (First Letter : Capital - Then:From 2 To 20 Any Capital Or Small Letters)`
+        productDesc.classList.add("is-invalid")
+        productDesc.classList.remove("is-valid")
+        alert4.classList.add("d-block")
+        alert4.innerHTML= `Enter Valid Description (First Letter : Capital - Then:From 2 To 20 Any Capital Or Small Letters)`
+
+        errors += `Enter Valid Description `
         return false
     }
 }
